pyproject.toml
[project]
name = "speaker-diarization"
version = "0.1.0"
description = "speaker-diarization"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "nemo-toolkit[asr]",
    "omegaconf>=2.3.0",
    "torch>=2.9.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
